Refactor floating-point calculations and update includes
- Changed `pow` to `powf` in `ConjGradOptimizer.cpp` for improved performance and precision with float types.
- Updated constants from `1.0` to `1.0f` for type consistency.
- Simplified calculation of `projNSSq` by removing unnecessary type casting.
- Modified `PSEUDO_EPS` in `Pseudoinverse.h` to use a direct float literal for clarity.
- Added includes and pragma directives in `HTMLDefines.h` to manage ATL/MFC warnings.
- Updated `theWheel.chm` with binary differences.

+++ src/OptimizeN/include/Pseudoinverse.h
@@ -11,7 +11,7 @@
 #include <MatrixNxM.h>
 
 // constant for pseudo-inverse
-const REAL PSEUDO_EPS = (REAL) 1e-8;
+const REAL PSEUDO_EPS = 1e-8f;
 
 //////////////////////////////////////////////////////////////////
 template<class TYPE>
